+++ docs/changelog.md
@@ -486,7 +486,7 @@
 - new `CommandFlow` and `ProcessExt` method: `command_shell`.
 
 ## yard
-- new macros: `_doc_meta!`, `_doc_size_of!`.
+- new macros: `_doc_meta!`, `_doc_test_size_of!`.
 - new hidden maintenance aliases for item attributes:
   - `__doc_<show|hide>`, `__cfg_item_[un]safe[_hide]`, `_js_[un]safe_ffi[_doc]`, `_unsafe_syscall_not_miri`.
 - update `_use_or_shim!` to add `_doc!` macro support.